+++ handler/order/delete.go
@@ -1,0 +1,23 @@
+package order
+
+import (
+	"github.com/gin-gonic/gin"
+	"net/http"
+	"shopping-cart/util"
+)
+
+func (h *Order) DeleteOrder(c *gin.Context) {
+	id, err := util.GetIDFromPath(c, "id")
+	if err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{"error": "invalid order ID"})
+		return
+	}
+
+	err = h.orderService.DeleteOrder(id)
+	if err != nil {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+		return
+	}
+
+	c.JSON(http.StatusOK, gin.H{"message": "Order deleted successfully"})
+}
