+++ handler/order/get.go
@@ -1,0 +1,33 @@
+package order
+
+import (
+	"github.com/gin-gonic/gin"
+	"net/http"
+	"shopping-cart/util"
+)
+
+func (h *Order) GetOrder(c *gin.Context) {
+	id, err := util.GetIDFromPath(c, "id")
+	if err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{"error": "invalid order ID"})
+		return
+	}
+
+	order, err := h.orderService.GetOrderByID(id)
+	if err != nil {
+		c.JSON(http.StatusNotFound, gin.H{"error": "Order not found"})
+		return
+	}
+
+	c.JSON(http.StatusOK, gin.H{"order": order})
+}
+
+func (h *Order) ListOrders(c *gin.Context) {
+	orders, err := h.orderService.ListAllOrders()
+	if err != nil {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": err})
+		return
+	}
+
+	c.JSON(http.StatusOK, gin.H{"orders": orders})
+}
