--- handler/order/order.go
@@ -1,90 +1,0 @@
-package order
-
-import (
-	"github.com/gin-gonic/gin"
-	"net/http"
-	"shopping-cart/model/datatransfer"
-	"shopping-cart/util"
-)
-
-func (h *Order) CreateOrder(c *gin.Context) {
-	var orderRequest datatransfer.OrderRequest
-	err := c.ShouldBindJSON(&orderRequest)
-	if err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
-		return
-	}
-
-	order, err := h.orderService.CreateOrder(&orderRequest)
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
-		return
-	}
-
-	c.JSON(http.StatusOK, gin.H{"message": "Order created successfully", "order": order})
-}
-
-func (h *Order) GetOrder(c *gin.Context) {
-	id, err := util.GetIDFromPath(c, "id")
-	if err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "invalid order ID"})
-		return
-	}
-
-	order, err := h.orderService.GetOrderByID(id)
-	if err != nil {
-		c.JSON(http.StatusNotFound, gin.H{"error": "Order not found"})
-		return
-	}
-
-	c.JSON(http.StatusOK, gin.H{"order": order})
-}
-
-func (h *Order) UpdateOrder(c *gin.Context) {
-	id, err := util.GetIDFromPath(c, "id")
-	if err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "invalid order ID"})
-		return
-	}
-
-	var orderRequest datatransfer.OrderRequest
-	err = c.ShouldBindJSON(&orderRequest)
-	if err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
-		return
-	}
-
-	order, err := h.orderService.UpdateOrder(id, &orderRequest)
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
-		return
-	}
-
-	c.JSON(http.StatusOK, gin.H{"message": "Order updated successfully", "order": order})
-}
-
-func (h *Order) DeleteOrder(c *gin.Context) {
-	id, err := util.GetIDFromPath(c, "id")
-	if err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "invalid order ID"})
-		return
-	}
-
-	err = h.orderService.DeleteOrder(id)
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
-		return
-	}
-
-	c.JSON(http.StatusOK, gin.H{"message": "Order deleted successfully"})
-}
-
-func (h *Order) ListOrders(c *gin.Context) {
-	orders, err := h.orderService.ListAllOrders()
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": err})
-		return
-	}
-
-	c.JSON(http.StatusOK, gin.H{"orders": orders})
-}
