+++ handler/order/update.go
@@ -1,0 +1,31 @@
+package order
+
+import (
+	"github.com/gin-gonic/gin"
+	"net/http"
+	"shopping-cart/model/datatransfer"
+	"shopping-cart/util"
+)
+
+func (h *Order) UpdateOrder(c *gin.Context) {
+	id, err := util.GetIDFromPath(c, "id")
+	if err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{"error": "invalid order ID"})
+		return
+	}
+
+	var orderRequest datatransfer.OrderRequest
+	err = c.ShouldBindJSON(&orderRequest)
+	if err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+		return
+	}
+
+	order, err := h.orderService.UpdateOrder(id, &orderRequest)
+	if err != nil {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+		return
+	}
+
+	c.JSON(http.StatusOK, gin.H{"message": "Order updated successfully", "order": order})
+}
